+++ CHANGELOG.md
@@ -1,6 +1,19 @@
-# PowerScp - Change History
+# PowerSCP - Development Change History
 
-## Version 0.1.0a (Development build) - 29.07.2020 
+## Version 0.2.1a (Development build) - 03.08.2020
 
-- Reorganized module struction and function into a single file
+- Added *-TransferFilesOnly* parameter to *Send-ScpItem* cmdlet
+- *Send-ScpItem* will now create remote destination directory if it does not exist
+- Cleaned up development comments and staging tree
+
+## Version 0.2.0a (Development build) - 31.07.2020
+
+- Added *Get-ScpItemType* cmdlet
+- Added *Test-ScpPath* cmdlet
+- Added Send-ScpItemcmdlet
+- Added *New-ScpDirectory* cmdlet
+
+## Version 0.1.0a (Development build) - 29.07.2020
+
+- Reorganized module functions into a single file
 - Created staging folder for all cmdlets/snippets not yet part of the main module files